Remove uneeded type mask

+++ src/spi.c
@@ -9,7 +9,7 @@ static uint32_t pin_to_bit_pos_conversion(unsigned int pin)
 void assert_spi_pin(uint32_t* gpio_output_addr, unsigned int gpio_pin)
 {
 	if (gpio_pin >= 16) { return; }
-	*gpio_output_addr |= (uint16_t) pin_to_bit_pos_conversion(gpio_pin);
+	*gpio_output_addr |= pin_to_bit_pos_conversion(gpio_pin);
 }
 
 void deassert_spi_pin(uint32_t* gpio_output_addr, unsigned int gpio_pin)
